pyproject.toml
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings"
]
markers = [
    "repositories: marks tests for repository layer",
    "api: marks tests for API layer",
    "helpers: marks tests for helper functions",
    "models: marks tests for models layer",
    "ruzapi: marks tests for ruz_api layer",
    "integration: marks integration tests",
    "unit: marks unit tests",
    "slow: marks tests as slow"
]

[tool.coverage.report]
fail_under = 70
show_missing = true

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "ruz-server"
version = "1.0.1"
description = "RUZ Server API"
dependencies = [
    "aiohttp==3.12.15",
    "apscheduler==3.11.2",
    "fastapi==0.135.2",
    "pydantic-settings==2.6.1",
    "sqlalchemy==2.0.41",
    "sqlmodel==0.0.24",
    "uvicorn==0.30.6",
    "httpx==0.28.1",
    "psycopg2==2.9.9",
]

[project.optional-dependencies]
test = [
    "pytest==8.3.3",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.1.0",
    "pytest-mock==3.15.1",
]
dev = [
    "ruff==0.15.10",
    "uv==0.11.6",
    "pip_audit==2.10.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"src/ruz_server/api/**/*.py" = ["B008"]
